Type : refactor Subject : 파일 구조 변경 및 일부 내용 수정
